Breaking changes: respect `--fileoutput` in CSV and rename `-f csv-export` to `-f csv` (#290)
# Motivation
We've had a number of bug reports over email about the csv export which
doesn't respect `--fileoutput`.

This PR fixes that by changing the csv export so it respects
`--fileoutput`. For any users who prefer the old behavior, I've added a
`--fileoutput-dynamic` flag that they can use to keep the previous
behavior.

# Other Changes
I've also renamed the csv export so it is consistent with other exporter
names. I've removed redundant code and made a tiny improvement to the
table output in response to a common source of confusion w/ users.

# Testing
I tested this PR manually by checking all the new flags and by checking
behavior without any flags to verify that everything works as expected.